create image and collection models, refactor blog
