+++ server/src/services/PermissionService.js
@@ -1,7 +1,7 @@
 /* eslint-disable consistent-return */
 /* eslint-disable no-return-await */
 const Permission = require('../models/Users/Permission.model')
-const { decamelize } = require('../helpers/decamelize')
+const { decamelize } = require('../helpers/utilities')
 
 module.exports = {
   async create(permission) {
